fixed e2e tests, added more integration and unit tests, modified home.html and auth styles

+++ src/models/Posts.model.js
@@ -576,8 +576,8 @@ module.exports.insertSaved = async function insertSaved(data) {
 
 // delete a save
 module.exports.deleteSavedByID = async function deleteSavedByID(data) {
-  const VALUES = [data.id];
-  const { rows } = await pool.query('DELETE FROM "SavedPosts" WHERE "id" = $1 RETURNING *', VALUES);
+  const VALUES = [data.id, data.user_id];
+  const { rows } = await pool.query('DELETE FROM "SavedPosts" WHERE "id" = $1 AND user_id = $2 RETURNING *', VALUES);
   return rows[0];
 };
 
@@ -732,7 +732,12 @@ module.exports.getSortedPosts = async function getSortedPosts(data) {
   }
 
   // Category filter
-  const categoryFilter = category ? `AND p.category = '${category}'` : '';
+  const params = [];
+  let categoryFilter = '';
+  if (category) {
+    params.push(category);
+    categoryFilter = `AND p.category = $${params.length}`;
+  }
 
   // Sort order
   const orderMap = {
